adding checks and updating the database architecture

+++ src/controllers/course_controllers.js
@@ -28,8 +28,8 @@ const getAllCourse = async (req, res) => {
         },
       ],
     });
-    if (result === null) {
-      return res.status(501).json({ message: "courses is null" });
+    if (result === null || result.length === 0) {
+      return res.status(404).json({ message: "courses is null" });
     } else {
       return res.status(200).json({ result });
     }
@@ -49,14 +49,14 @@ const addCourse = async (req, res) => {
   const { id, title } = req.body;
 
   try {
-    if (id === 1) {
-      return res.status(501).json({ message: "wrong body" });
+    if (id === null || title === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
       const result = await Courses.create({
         admin_id: id,
         title: title,
       });
-      if (result === null) {
+      if (result === null || result.length === 0) {
         console.error(
           "Ошибка при выполнение запроса при добавления курса (res null)",
           error
@@ -80,17 +80,24 @@ const updateTitleCourse = async (req, res) => {
   const { id, admin_id, title } = req.body;
 
   try {
-    const result = await Courses.update({ title }, { where: { id, admin_id } });
-    if (result === null) {
-      console.error(
-        "Ошибка при выполнение запроса на обновление названия курса (res null)",
-        error
-      );
-      return res
-        .status(501)
-        .json({ message: "Ошибка при обновление названия курса" });
+    if (title === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await Courses.update(
+        { title },
+        { where: { id, admin_id } }
+      );
+      if (result === null || result.length === 0) {
+        console.error(
+          "Ошибка при выполнение запроса на обновление названия курса (res null)",
+          error
+        );
+        return res
+          .status(501)
+          .json({ message: "Ошибка при обновление названия курса" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -109,21 +116,25 @@ const addUserInCourse = async (req, res) => {
   const { user_id, course_id } = req.body;
 
   try {
-    const result = await CoursesOfUsers.create({
-      user_id: user_id,
-      course_id: course_id,
-      auth_in_course: false,
-    });
-    if (result === null) {
-      console.error(
-        "Ошибка при выполнение запроса при добавления пользователя на курс (res null)",
-        error
-      );
-      return res
-        .status(501)
-        .json({ message: "Ошибка при добавления пользователя на курс" });
+    if (user_id === null || course_id === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await CoursesOfUsers.create({
+        user_id: user_id,
+        course_id: course_id,
+        auth_in_course: false,
+      });
+      if (result === null || result.length === 0) {
+        console.error(
+          "Ошибка при выполнение запроса при добавления пользователя на курс (res null)",
+          error
+        );
+        return res
+          .status(501)
+          .json({ message: "Ошибка при добавления пользователя на курс" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -143,20 +154,24 @@ const addThemeInCourse = async (req, res) => {
   const { title, course_id } = req.body;
 
   try {
-    const result = await ThemesOfCourses.create({
-      title: title,
-      course_id: course_id,
-    });
-    if (result === null) {
-      console.error(
-        "Ошибка при выполнение запроса при добавления темы на курс (res null)",
-        error
-      );
-      return res
-        .status(501)
-        .json({ message: "Ошибка при добавления темы на курс" });
+    if (title === null || course_id === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await ThemesOfCourses.create({
+        title: title,
+        course_id: course_id,
+      });
+      if (result === null || result.length === 0) {
+        console.error(
+          "Ошибка при выполнение запроса при добавления темы на курс (res null)",
+          error
+        );
+        return res
+          .status(501)
+          .json({ message: "Ошибка при добавления темы на курс" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -174,20 +189,24 @@ const addThemeInCourse = async (req, res) => {
 const getAllThemesInCourses = async (req, res) => {
   const { course_id } = req.body;
   try {
-    const result = await ThemesOfCourses.findAll({
-      where: { course_id },
-      attributes: ["id", "title", "course_id"],
-      include: [
-        {
-          model: Courses,
-          attributes: ["title"],
-        },
-      ],
-    });
-    if (result === null) {
-      return res.status(501).json({ message: "themes is null" });
+    if (course_id === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await ThemesOfCourses.findAll({
+        where: { course_id },
+        attributes: ["id", "title", "course_id"],
+        include: [
+          {
+            model: Courses,
+            attributes: ["title"],
+          },
+        ],
+      });
+      if (result === null || result.length === 0) {
+        return res.status(501).json({ message: "themes is null" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -204,21 +223,25 @@ const getAllThemesInCourses = async (req, res) => {
 const getAllTUsersInCourses = async (req, res) => {
   const { course_id } = req.body;
   try {
-    const result = await CoursesOfUsers.findAll({
-      where: { course_id },
-      attributes: ["id", "user_id", "course_id", "auth_in_course"],
-      include: [
-        {
-          model: Users,
-          as: "user",
-          attributes: ["fio", "isAdmin", "isAuth"],
-        },
-      ],
-    });
-    if (result === null) {
-      return res.status(501).json({ message: "users is null" });
+    if (course_id === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await CoursesOfUsers.findAll({
+        where: { course_id },
+        attributes: ["id", "user_id", "course_id", "auth_in_course"],
+        include: [
+          {
+            model: Users,
+            as: "user",
+            attributes: ["fio", "isAdmin", "isAuth"],
+          },
+        ],
+      });
+      if (result === null || result.length === 0) {
+        return res.status(501).json({ message: "users is null" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -235,17 +258,21 @@ const getAllTUsersInCourses = async (req, res) => {
 const updateThemeInCourse = async (req, res) => {
   const { id, title } = req.body;
   try {
-    const result = await ThemesOfCourses.update({ title }, { where: { id } });
-    if (result === null) {
-      console.error(
-        "Ошибка при выполнение запроса на обновление данных темы (res null)",
-        error
-      );
-      return res
-        .status(501)
-        .json({ message: "Ошибка при обновление данных о теме" });
+    if (id === null || title === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await ThemesOfCourses.update({ title }, { where: { id } });
+      if (result === null || result.length === 0) {
+        console.error(
+          "Ошибка при выполнение запроса на обновление данных темы (res null)",
+          error
+        );
+        return res
+          .status(501)
+          .json({ message: "Ошибка при обновление данных о теме" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -263,20 +290,24 @@ const updateThemeInCourse = async (req, res) => {
 const changeAuthUserInCourse = async (req, res) => {
   const { id, auth_in_course } = req.body;
   try {
-    const result = await CoursesOfUsers.update(
-      { auth_in_course },
-      { where: { id } }
-    );
-    if (result === null) {
-      console.error(
-        "Ошибка при выполнение запроса на обновление данных auth юзера в теме (res null)",
-        error
-      );
-      return res
-        .status(501)
-        .json({ message: "Ошибка при обновление данных об auth юзера в теме" });
+    if (id === null || auth_in_course === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await CoursesOfUsers.update(
+        { auth_in_course },
+        { where: { id } }
+      );
+      if (result === null || result.length === 0) {
+        console.error(
+          "Ошибка при выполнение запроса на обновление данных auth юзера в теме (res null)",
+          error
+        );
+        return res.status(501).json({
+          message: "Ошибка при обновление данных об auth юзера в теме",
+        });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -295,15 +326,19 @@ const changeAuthUserInCourse = async (req, res) => {
 const checkThemeIsBusy = async (req, res) => {
   const { id } = req.body;
   try {
-    const relatedRecords = await Promise.all([
-      QuestionsOfThemes.count({ where: { theme_id: id } }),
-      Schedule.count({ where: { theme_id: id } }),
-      Progress.count({ where: { theme_id: id } }),
-    ]);
+    if (id === null) {
+      return res.status(501).json({ message: "body is null" });
+    } else {
+      const relatedRecords = await Promise.all([
+        QuestionsOfThemes.count({ where: { theme_id: id } }),
+        Schedule.count({ where: { theme_id: id } }),
+        Progress.count({ where: { theme_id: id } }),
+      ]);
 
-    const isBusy = relatedRecords.some((count) => count > 0);
+      const isBusy = relatedRecords.some((count) => count > 0);
 
-    return res.status(200).json({ result: isBusy });
+      return res.status(200).json({ result: isBusy });
+    }
   } catch (error) {
     console.error("Ошибка при выполнение запроса на занятость темы", error);
     res.status(500).json({
@@ -317,17 +352,21 @@ const checkThemeIsBusy = async (req, res) => {
 const deleteTheme = async (req, res) => {
   const { id } = req.body;
   try {
-    await QuestionsOfThemes.destroy({ where: { theme_id: id } });
-    await Schedule.destroy({ where: { theme_id: id } });
-    await Progress.destroy({ where: { theme_id: id } });
+    if (id === null) {
+      return res.status(501).json({ message: "body is null" });
+    } else {
+      await QuestionsOfThemes.destroy({ where: { theme_id: id } });
+      await Schedule.destroy({ where: { theme_id: id } });
+      await Progress.destroy({ where: { theme_id: id } });
 
-    const result = await ThemesOfCourses.destroy({
-      where: {
-        id: id,
-      },
-    });
+      const result = await ThemesOfCourses.destroy({
+        where: {
+          id: id,
+        },
+      });
 
-    return res.status(200).json({ result });
+      return res.status(200).json({ result });
+    }
   } catch (error) {
     console.error("Ошибка при выполнение запроса на занятость темы", error);
     res.status(500).json({
@@ -341,20 +380,24 @@ const deleteTheme = async (req, res) => {
 const getAllQuestionsOfThemes = async (req, res) => {
   const { id } = req.body;
   try {
-    const result = await QuestionsOfThemes.findAll({
-      where: { id },
-      attributes: ["id", "title", "theme_id"],
-      include: [
-        {
-          model: ThemesOfCourses,
-          attributes: ["title"],
-        },
-      ],
-    });
-    if (result === null) {
-      return res.status(501).json({ message: "question is null" });
+    if (id === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await QuestionsOfThemes.findAll({
+        where: { id },
+        attributes: ["id", "title", "theme_id"],
+        include: [
+          {
+            model: ThemesOfCourses,
+            attributes: ["title"],
+          },
+        ],
+      });
+      if (result === null || result.length === 0) {
+        return res.status(501).json({ message: "question is null" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -371,17 +414,24 @@ const getAllQuestionsOfThemes = async (req, res) => {
 const updateQuestionInTheme = async (req, res) => {
   const { id, title } = req.body;
   try {
-    const result = await QuestionsOfThemes.update({ title }, { where: { id } });
-    if (result === null) {
-      console.error(
-        "Ошибка при выполнение запроса на обновление данных вопроса (res null)",
-        error
-      );
-      return res
-        .status(501)
-        .json({ message: "Ошибка при обновление данных о вопросе" });
+    if (id === null || title === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ result });
+      const result = await QuestionsOfThemes.update(
+        { title },
+        { where: { id } }
+      );
+      if (result === null || result.length === 0) {
+        console.error(
+          "Ошибка при выполнение запроса на обновление данных вопроса (res null)",
+          error
+        );
+        return res
+          .status(501)
+          .json({ message: "Ошибка при обновление данных о вопросе" });
+      } else {
+        return res.status(200).json({ result });
+      }
     }
   } catch (error) {
     console.error(
@@ -407,7 +457,7 @@ const deleteQuestion = async (req, res) => {
           id: id,
         },
       });
-      if (result === null) {
+      if (result === null || result.length === 0) {
         console.error(
           "Ошибка при выполнение запроса на удаление вопроса (res null)",
           error
@@ -432,22 +482,26 @@ const addQuestionInTheme = async (req, res) => {
   }
 
   try {
-    const createPromises = title.map(async (t) => {
-      return await QuestionsOfThemes.create({
-        title: t,
-        theme_id: theme_id,
-      });
-    });
-
-    const results = await Promise.all(createPromises);
-
-    if (results.some((result) => result === null)) {
-      console.error("Ошибка при создании некоторых вопросов");
-      return res
-        .status(501)
-        .json({ message: "Ошибка при добавлении некоторых вопросов на тему" });
+    if (theme_id === null) {
+      return res.status(501).json({ message: "body is null" });
     } else {
-      return res.status(200).json({ results });
+      const createPromises = title.map(async (t) => {
+        return await QuestionsOfThemes.create({
+          title: t,
+          theme_id: theme_id,
+        });
+      });
+
+      const results = await Promise.all(createPromises);
+
+      if (results.some((result) => result === null || result.length === 0)) {
+        console.error("Ошибка при создании некоторых вопросов");
+        return res.status(501).json({
+          message: "Ошибка при добавлении некоторых вопросов на тему",
+        });
+      } else {
+        return res.status(200).json({ results });
+      }
     }
   } catch (error) {
     console.error(
